Enable and update regression test
